use service client rather than azServiceClient
